+++ R/Lib_PROSAIL_HybridInversion.R
@@ -41,6 +41,42 @@ apply_noise_atbd <- function(BRF_LUT){
   return(BRF_LUT_Noise)
 }
 
+#' This function applies additive and multiplicative noise to BRF data
+#' - if AdditiveNoise and MultiplicativeNoise are defined with a unique value,
+#' noise is homogeneous across all spectrum
+#' - if AdditiveNoise and MultiplicativeNoise are the same length as the number
+#' of spectral bands (rows in BRF_LUT), noise is specific to each spectral band
+#'
+#' @param BRF_LUT numeric. BRF LUT
+#' @param AdditiveNoise numeric. additive noise (0 = 0%, 1 = 100%)
+#' @param MultiplicativeNoise numeric. multiplicative noise (0 = 0%, 1 = 100%)
+#'
+#' @return BRF_LUT_Noise numeric.
+#' @export
+
+apply_noise_AddMult <- function(BRF_LUT,
+                                AdditiveNoise = 0.01,
+                                MultiplicativeNoise = 0.02){
+  nbWL <- nrow(BRF_LUT)
+  nbSamples <- ncol(BRF_LUT)
+  # add noise to BRF
+  if (length(AdditiveNoise)==1){
+    AddComp <- matrix(rnorm(nbWL*nbSamples,0,AdditiveNoise),
+                      nrow = nbWL)
+  } else if ((length(AdditiveNoise)==nbWL)){
+    AddComp <- matrix(data = 0, nrow = nbWL, ncol = nbSamples)
+    for (i in 1:nbWL) AddComp[i,] <- matrix(data = rnorm(nbSamples,0,AdditiveNoise[i]), ncol = nbSamples)
+  }
+  if (length(MultiplicativeNoise)==1){
+    MultComp <- matrix(rnorm(nbWL*nbSamples,0,MultiplicativeNoise), nrow = nbWL)
+  } else if ((length(MultiplicativeNoise)==nbWL)){
+    MultComp <- matrix(data = 0, nrow = nbWL, ncol = nbSamples)
+    for (i in 1:nbWL) MultComp[i,] <- matrix(data = rnorm(nbSamples,0,MultiplicativeNoise[i]),ncol = nbSamples)
+  }
+  BRF_LUT_Noise <- BRF_LUT*(1+(MultComp)) + AddComp
+  return(BRF_LUT_Noise)
+}
+
 
 #' This function applies SVR model on raster data in order to estimate
 #' vegetation biophysical properties
@@ -641,11 +677,23 @@ train_prosail_inversion <- function(InputPROSAIL = NULL, BRF_LUT = NULL,
     # check if same spectral sampling for all key variables
     check_SpectralSampling(SpecPROSPECT, SpecSOIL, SpecATM)
     # generate LUT of BRF corresponding to InputPROSAIL, for a sensor
-    BRF_LUT <- Generate_LUT_BRF(SAILversion = SAILversion,
-                                InputPROSAIL = InputPROSAIL,
-                                SpecPROSPECT = SpecPROSPECT,
-                                SpecSOIL = SpecSOIL,
-                                SpecATM = SpecATM)
+    if (length(which(!is.na(match(c('fCover', 'albedo', 'fAPAR'), Parms2Estimate))))==0){
+      BRF_LUT <- Generate_LUT_BRF(SAILversion = SAILversion,
+                                  InputPROSAIL = InputPROSAIL,
+                                  SpecPROSPECT = SpecPROSPECT,
+                                  SpecSOIL = SpecSOIL,
+                                  SpecATM = SpecATM)
+    } else if (length(which(!is.na(match(c('fCover', 'albedo', 'fAPAR'), Parms2Estimate))))>0){
+      res <- Generate_LUT_PROSAIL(SAILversion = SAILversion,
+                                  InputPROSAIL = InputPROSAIL,
+                                  SpecPROSPECT = SpecPROSPECT,
+                                  SpecSOIL = SpecSOIL,
+                                  SpecATM = SpecATM)
+      BRF_LUT <- res$BRF
+      InputPROSAIL$fCover <- res$fCover
+      InputPROSAIL$fAPAR <- res$fAPAR
+      InputPROSAIL$albedo <- res$albedo
+    }
 
     ### == == == == == == == == == == == == == == == == == == == == == == ###
     ###     3- APPLY SPECTRAL RESPONSE FUNCTION if not already applied    ###